Fix compiler warning with strlen

+++ projects/win32/ignition_server/driver_log.cpp
@@ -5,7 +5,7 @@
 namespace ignition {
 
   void DriverLog::Log(const char *pchLogMessage) {
-    int logLen = strlen(pchLogMessage);
+    size_t logLen = strlen(pchLogMessage);
     bool hasNewline = pchLogMessage[logLen-1] == '\n';
     printf("%s%s", pchLogMessage, hasNewline ? "" : "\n");
   }
